requirements.txt
numpy~=1.24.3
torchvision~=0.18.1
pillow~=10.4.0
setuptools~=69.5.1